1/9 기능수정 1. TowerHp, TowerMana 스크립터블오브젝트로 쓰던것을 없애고 gameHUD에 있는것으로 대체함 UIManager에서 gameHUD에 접근하고 있으므로 UIManager를 통해 접근하게 됨

+++ TowerDefense/Assets/Scripts/DataControl/TowerDataControl/ManaUsingTowerData.cs
@@ -6,8 +6,8 @@ namespace DataControl.TowerDataControl
     [CreateAssetMenu(menuName = "Tower Data/Attack Tower Data/Mana Tower Data")]
     public class ManaUsingTowerData : AttackTowerData
     {
-        public float attackMana { get; private set; }
-        [SerializeField] private int initMana;
+        public byte attackMana { get; private set; }
+        [SerializeField] private byte initMana;
 
         public override void InitState()
         {
